OnlineShop/requirements.txt
Django 
djangorestframework
djangorestframework-simplejwt
drf-yasg
PyJWT
django-filter
django-cors-headers


requests
pgvector


channels
channels-redis


react
